fix: added glue code to return array for tests which are expecting it - temp fix, added an issue to track #152
