+++ src/store/matchedAudiences.ts
@@ -1,46 +1,23 @@
 import { storage, timeStampInSecs } from '../utils';
-import { StorageKeys, MatchedAudience, AudienceDefinition } from '../../types';
-
-type MatchedVersionLookup = {
-  [key: string]: number
-}
+import {
+  StorageKeys,
+  MatchedAudience,
+  MatchedAudiences,
+  AudienceDefinition,
+} from '../../types';
 
 class MatchedAudienceStore {
-  private matchedAudiences: MatchedAudience[];
+  private matchedAudiences: MatchedAudiences;
   private matchedAudienceIds: string[];
-  private matchedVersionLookup: MatchedVersionLookup;
+  private unsetDueToVersionIncAudienceIds: string[];
+  private storeLoadedAt: number;
 
   constructor() {
-    this.matchedAudiences = [];
+    this.matchedAudiences = {};
     this.matchedAudienceIds = [];
-    this.matchedVersionLookup = {}
+    this.unsetDueToVersionIncAudienceIds = [];
+    this.storeLoadedAt = timeStampInSecs();
     this._load();
-  }
-
-  _createMatchedVersionLookup(audiences: MatchedAudience[]): MatchedVersionLookup {
-    return audiences.reduce((acc, cur) => ({
-      ...acc, [cur.id]: cur.version
-    }), {});
-  }
-
-  _load(): void {
-    const audiences: MatchedAudience[] =
-      storage.get(StorageKeys.MATCHED_AUDIENCES) || [];
-    const unExpiredAudiences = audiences
-      .filter((audience) => audience.expiresAt > timeStampInSecs())
-      .map((audience) => {
-        return {
-          ...audience,
-          matchedOnCurrentPageView: false,
-        };
-      });
-    const unExpiredAudienceIds = unExpiredAudiences.map(
-      (audience) => audience.id
-    );
-    this.matchedAudiences = unExpiredAudiences;
-    this.matchedAudienceIds = unExpiredAudienceIds;
-    this.matchedVersionLookup = this._createMatchedVersionLookup(unExpiredAudiences);
-    this._save();
   }
 
   _save(): void {
@@ -48,39 +25,81 @@ class MatchedAudienceStore {
     storage.set(StorageKeys.MATCHED_AUDIENCE_IDS, this.matchedAudienceIds);
   }
 
-  hasAudienceBeenMatched(audienceId: string, audienceVersion: number): boolean {
-    if (this.matchedVersionLookup[audienceId]) {
-      if (this.matchedVersionLookup[audienceId] >= audienceVersion) {
-        return true;
-      }
+  _hasAudienceExpired(expiresAt: number): boolean {
+    return expiresAt < this.storeLoadedAt;
+  }
+
+  _unsetAudience(id: string): void {
+    delete this.matchedAudiences[id];
+  }
+
+  _updatePageViewFlag(id: string, state: boolean): void {
+    this.matchedAudiences[id].matchedOnCurrentPageView = state;
+  }
+
+  _load(): void {
+    // TODO: @ydennisy remove this backward compat code.
+    // https://github.com/AirGrid/edgekit/issues/152
+    let loadedAudiences: MatchedAudiences =
+      storage.get(StorageKeys.MATCHED_AUDIENCES) || {};
+
+    if (Array.isArray(loadedAudiences)) {
+      storage.remove(StorageKeys.MATCHED_AUDIENCES);
+      loadedAudiences = loadedAudiences.reduce((acc, cur) => {
+        acc[cur.id] = cur;
+      }, {});
     }
-    return false;
-  }
 
-  unmatchAudiencesDueToVersionUpdate(audienceDefinitions: AudienceDefinition[]): void {
-    audienceDefinitions.forEach((audience) => {
-      
-    })
-  }
+    this.matchedAudiences = loadedAudiences;
 
-  setMatchedAudiences(matchedAudiences: MatchedAudience[]): void {
-    const audiencesMatchedOnCurrentLoad = matchedAudiences.map((audience) => {
-      if (this.matchedVersionLookup[audience.id]) {
-        audience.matchedOnCurrentPageView = false;
+    Object.entries(loadedAudiences).forEach(([id, audience]) => {
+      if (this._hasAudienceExpired(audience.expiresAt)) {
+        this._unsetAudience(id);
       }
-      return audience
+      this._updatePageViewFlag(id, false);
     });
-    this.matchedAudiences = [
-      ...this.matchedAudiences, 
-      ...audiencesMatchedOnCurrentLoad
-    ];
-    this.matchedVersionLookup = this._createMatchedVersionLookup(this.matchedAudiences);
-    this.matchedAudienceIds = this.matchedAudiences.map((audience) => audience.id);
+
+    this.matchedAudienceIds = Object.keys(this.matchedAudiences);
     this._save();
   }
 
-  getMatchedAudiences(): MatchedAudience[] {
-    return [...this.matchedAudiences];
+  unsetAudiencesIfVersionIncremented(
+    audienceDefinitions: AudienceDefinition[]
+  ): void {
+    audienceDefinitions.forEach((audience) => {
+      const incomingVersion = audience.version;
+      const currentVersion = this.matchedAudiences[audience.id]
+        ? this.matchedAudiences[audience.id].version
+        : null;
+      if (currentVersion && currentVersion < incomingVersion) {
+        this._unsetAudience(audience.id);
+        this.unsetDueToVersionIncAudienceIds.push(audience.id);
+      }
+    });
+  }
+
+  isMatched(id: string, version: number): boolean {
+    return !!(
+      this.matchedAudiences[id] && this.matchedAudiences[id].version >= version
+    );
+  }
+
+  setAudiences(matchedAudiences: MatchedAudience[]) {
+    matchedAudiences.forEach((audience) => {
+      this.matchedAudiences[audience.id] = audience;
+      if (this.unsetDueToVersionIncAudienceIds.includes(audience.id)) {
+        this._updatePageViewFlag(audience.id, false);
+      }
+    });
+    this._save();
+  }
+
+  getMatchedAudiences() {
+    // TODO: this is for backward compat.
+    // https://github.com/AirGrid/edgekit/issues/152
+    return Object.entries(this.matchedAudiences).map(
+      ([_, audience]) => audience
+    );
   }
 }
