+++ test/e2e/run.test.ts
@@ -7,8 +7,13 @@ type Store = { edkt_matched_audiences: string; edkt_page_views: string };
 
 const getPageViewsFromStore = (store: Store) =>
   JSON.parse(store['edkt_page_views']);
-const getMatchedAudiencesFromStore = (store: Store) =>
-  JSON.parse(store['edkt_matched_audiences']);
+const getMatchedAudiencesFromStore = (store: Store) => {
+  // TODO: this is added for backward compat.
+  // https://github.com/AirGrid/edgekit/issues/152
+  const matchedAudiences = JSON.parse(store['edkt_matched_audiences']);
+  return Object.entries(matchedAudiences).map(([_, audience]) => audience);
+}
+
 const getLocalStorageFromPage = (): Promise<Store> =>
   page.evaluate('localStorage');
