+++ test/helpers/localStorage.ts
@@ -1,8 +1,7 @@
 import {
   PageView,
-  AudienceDefinition,
-  CachedAudienceMetaData,
   MatchedAudience,
+  MatchedAudiences,
 } from '../../types';
 import { viewStore, matchedAudienceStore } from '../../src/store';
 
@@ -16,7 +15,7 @@ export const clearStore = (): void => {
 export const setUpLocalStorage = (pageViews: PageView[]): void => {
   localStorage.clear();
   localStorage.setItem('edkt_page_views', JSON.stringify(pageViews));
-  //We need to reload from local storage because its only done on construction
+  // We need to reload from local storage because its only done on construction
   viewStore._load();
   matchedAudienceStore._load();
 };
@@ -24,11 +23,11 @@ export const setUpLocalStorage = (pageViews: PageView[]): void => {
 export const getPageViews = (): PageView[] =>
   JSON.parse(localStorage.getItem('edkt_page_views') || '[]');
 
-export const getMatchedAudiences = (): MatchedAudience[] =>
-  JSON.parse(localStorage.getItem('edkt_matched_audiences') || '[]');
-
-export const getCachedAudiences = (): AudienceDefinition[] =>
-  JSON.parse(localStorage.getItem('edkt_cached_audiences') || '[]');
-
-export const getCachedAudiencesMetaData = (): CachedAudienceMetaData =>
-  JSON.parse(localStorage.getItem('edkt_cached_audience_meta_data') || '{}');
+export const getMatchedAudiences = (): MatchedAudience[] => {
+  const matchedAudiences: MatchedAudiences = JSON.parse(
+    localStorage.getItem('edkt_matched_audiences') || '{}'
+  );
+  // TODO: this code has been added for backward compat
+  // https://github.com/AirGrid/edgekit/issues/152
+  return Object.entries(matchedAudiences).map(([_, audience]) => audience);
+};
