+++ test/unit/store/matchedAudiences.test.ts
@@ -1,5 +1,4 @@
 import { matchedAudienceStore } from '../../../src/store/matchedAudiences';
-//import { timeStampInSecs } from '../../../src/utils/index';
 import { clearStore } from '../../helpers/localStorage';
 import { MatchedAudience } from '../../../types';
 
@@ -10,9 +9,8 @@ const createMatchedAudience = (): MatchedAudience => {
     matchedAt: 1,
     expiresAt: 2,
     matchedOnCurrentPageView: true,
-  }
-}
-
+  };
+};
 
 describe('matchedAudienceStore', () => {
   beforeAll(clearStore);
@@ -21,32 +19,41 @@ describe('matchedAudienceStore', () => {
 
   it('should load an empty store', () => {
     const matchedAudiences = matchedAudienceStore.getMatchedAudiences();
-    
+
     expect(matchedAudiences).toHaveLength(0);
-  })
+  });
 
   it('should load and return a matched audience', () => {
-    matchedAudienceStore.setMatchedAudiences([matchedAudience])
+    matchedAudienceStore.setAudiences([matchedAudience]);
     const matchedAudiences = matchedAudienceStore.getMatchedAudiences();
 
     expect(matchedAudiences).toHaveLength(1);
-  })
+  });
 
   it('hasAudienceBeenMatched should return true, for a previously matched audience', () => {
-    const wasMatched = matchedAudienceStore.hasAudienceBeenMatched(matchedAudience.id, matchedAudience.version)
+    const wasMatched = matchedAudienceStore.isMatched(
+      matchedAudience.id,
+      matchedAudience.version
+    );
 
     expect(wasMatched).toBe(true);
-  })
+  });
 
   it('hasAudienceBeenMatched should return false, for a new version', () => {
-    const wasMatched = matchedAudienceStore.hasAudienceBeenMatched(matchedAudience.id, 999)
+    const wasMatched = matchedAudienceStore.isMatched(
+      matchedAudience.id,
+      999
+    );
 
     expect(wasMatched).toBe(false);
-  })
+  });
 
   it('hasAudienceBeenMatched should return false, for a new audience id', () => {
-    const wasMatched = matchedAudienceStore.hasAudienceBeenMatched('new', matchedAudience.version)
+    const wasMatched = matchedAudienceStore.isMatched(
+      'new',
+      matchedAudience.version
+    );
 
     expect(wasMatched).toBe(false);
-  })
-})
+  });
+});
